@@ -6,25 +6,37 @@
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
     "@mui/material": "^6.1.10",
+    "@tensorflow-models/handpose": "^0.1.0",
+    "@tensorflow/tfjs-backend-webgl": "^4.22.0",
+    "@tensorflow/tfjs-core": "^4.22.0",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^13.0.0",
     "@testing-library/user-event": "^13.2.1",
     "agora-rtc-sdk-ng": "^4.22.2",
     "axios": "^1.7.9",
+    "buffer": "^6.0.3",
     "cra-template": "1.2.0",
+    "crypto-browserify": "^3.12.1",
     "firebase": "^11.0.2",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "heapdump": "^0.3.15",
+    "moment": "^2.30.1",
+    "opencv.js": "^1.2.1",
+    "razorpay": "^2.9.5",
+    "react": "^19.0.0",
+    "react-big-calendar": "^1.16.3",
+    "react-datepicker": "^7.5.0",
+    "react-dom": "^19.0.0",
     "react-firebase-hooks": "^5.1.1",
     "react-router-dom": "^6.28.0",
     "react-scripts": "5.0.1",
+    "stream-browserify": "^3.0.0",
     "twilio-video": "^2.29.0",
     "web-vitals": "^2.1.0"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "test": "react-app-rewired test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -44,5 +56,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "react-app-rewired": "^2.2.1"
   }
 }
