@@ -4,7 +4,7 @@ import './Footer.css';
 const Footer = () => {
   return (
     <footer>
-      <p>&copy; 2024 Service Platform. All rights reserved.</p>
+      <p>&copy; 2024 TaskMaster. All rights reserved.</p>
     </footer>
   );
 };
